test_auto/requirements.txt
selenium
Appium-Python-Client
xmlrunner